New store package with AddCharm and OpenCharm interface.
The interface to the package is trivial, but internally it actually
handles all the necessary logic for concurrent runs of the algorithm,
including mongo-based atomic locks with expiration, multi-URL synchronous
revision bumping as described in the charm specification, GridFS-based
memory-friendly uploading for large files, and ponies too.

Lacks documentation and sha256 handling, though.. but I need some sleep.