store: do not log events internally on Publish
The external code that calls Publish has to deal with event logging
anyway for logging failures and whatnot. It's trivial for it to log the
error and success case from Publish too, so there's no reason to have
that logic duplicated in two different places. In fact, it's harder to
deal with it because the call site has to take care not to log errors
twice in that particular case.