Implement the publish API.
Use the proper permission checks when uploading charms or bundle.
Allow uploading and publishing at the same time.

This branch should complete the publishing work.
Later we will add promulgation policy as requested by eco.
In essence, on promulgating we can set the published write ACLs for
the entity to promulgators only, so that we don't allow
user publishing of promugated charms by default.