internal/v4: implement custom ResolveURL
This ensures that (most) ids returned by the v4 API
will have a series, even for multiple-series charms.

Also pave the way for v4 to be able to override specific
methods in the v5 router (for example, we want search-like
methods to be expand multi-series charms into multiple
result entities).

This required a bunch of shuffling around in internal/v5;
we took this opportunity to factor the various function arguments
to router.New into an interface, and removed the now-redundant
EntityExists function.