internal/entitycache: better tests
We make the entitycache test mocks more faithful to the actual
mgo primitives by making sure we always select only the fields
that have actually been selected.

We also add a test that promulgated URLs do not issue
a concurrent base-entity fetch (because they can't)
and ensure that the fields needed for the entitycache
logic are always fetched.