internal/charmstore: move upload logic from API level
The API tests were second-guessing the charmstore behaviour
with respect to blob uploading, which was fine but we're
changing the upload behaviour, so now we make all the
tests use the regular streaming blob upload code.

There are no semantic changes to the production code.

We will need to move some tests from the API levels
to the charmstore level, because charmstore package
coverage is now poor, but we will do that in a new PR
because this one is large enough already.

To summarise:

- the charm upload streaming code moves from internal/v5 to internal/charmstore.
- the code related to adding entities in internal/charmstore moves into addentity.go.
- in internal/storetesting we add NewCharm and NewBundle to make it easy to
generate a charm or bundle along with its archive.
- we remove relationTestingCharm and replace it with a more general variant that
uses storetesting.NewCharm.
- we fix a bug in charmstore.Pool.Close that meant the audit logger was never
closed (caught by go vet)
- we add a test helper that adds any charms required by a bundle, because
bundle verification now requires that all bundles added by tests have their
prerequisite charms in the store.
- we align the v4 tests somewhat closer with the diverged v5 tests.