internal/{v4, legacy}: serve metadata.yaml with empty series
Multi-series charms contain a list of supported series in the
"series" field in the metadata.yaml file. This field has changed
type, so some juju versions prior to 1.26 cannot parse the string
list form. To ensure compatibility with old version without
changing the name of the field, in legacy and v4 API versions,
we elide the series field from multi-series charms (for those
API versions, the series is always available in the charm id).