internal/charmstore: better migration testing, first step
We need better confidence in our migrations, so this
implements the start of a different approach to
migration testing.

It implements code to go back to historical revisions,
check out the code there and run the charm store at that
revision (including all the migrations that that revision
implies) and apply some updates. After that, the database
can be dumped.

This is triggered by the -dump-migration-history
test flag, which should only need to be invoked occasionally
(when new migrations are added on top of migrations
that are already running in the production server).

By default, the dumped database is restored and migrations
run on that.